Some work toward Python 3 compatibility.
This does some basic work to make at least most of this compatible
(importable, and valid syntax) with Python 3.

This also fix some weird mixing of tab and space, that may make things
look in a block on some machine, or outside of a block on some others.

With this (and a couple of other modification) I was able to get sphinx
to accept to build the documentation automatically.

Also Python2 End of Life is slightly over a year, so it will become
likely increasingly difficult soon after that to run install Python 2 on
new machines, so better get ready.