Set error code when existing prematurely from program.
The unix convention is to return with a non-zero error-code when
something was wrong, with the value of the error code which depends on
the reason for exist.

Especially in unit-test framework, the return code is critical.
This also affect the behavior of shells, like && and || in bash.

Hence assign an error code to some of the sys.exit

Note that in python you can also pass a string to sys.exit(...) and that
python will by itself return a non-zero error code, as well as print a
message.