Example Crosslink documentation with Dendropy.
This show the basic on how sphinx can be used to crosslink
documentation. Here it's a bit verbose, but usually something like

   :any:`DnaCharacterMatrix`

Should be sufficient to crosslink once the inventory are set.
Once the documentation is built, you should be able to see that the
`DnaCharacterMatrix` item in the rendered doc is a link to dendropy
documentation