Update some of the chdir.
In many cases instead of updating the cwd of the process we can just
pass in which cwd we want to run a given command. This appear to be a
bit cleaner.

There are some small distinction between subprocess.call and os.system,
the former is prefered. The `shell=True` parameter will do the argument
parsing in the subshell. We could also contruct the args manually and
not pass shell=True. That would be preferable in particular if file
names could have spaces in them.