Run tests in parallel and collect coverage informations.
This speeds up test by running them in parallel (including slow test on
my machine this takes ~3 minutes).

It also start collecting coverage information so that we can upload them
to codecov later and show the duration on each test.

So far we autodetect the number of cores to run the test so it should
also speed travis ci that gives us 2 cores.